Uploaded Lukasarts Tracer
You shouldn't have this.
This illegal piece of software uses speakers to transform mechanical waves into propagated EM waves. The EM wavelengths and frequencies are quadratic, designed to cause effects and molecular reactions in chemistry; remove the ATP process in cells, transform oxygen into nutritious oxide, and inflict hundreds of different symptoms to the human body from just a speaker. Just kidding, the program is fake. Have fun with it!